+++ src/url.hpp
@@ -9,15 +9,25 @@ using namespace geode::prelude;
 
 static std::map<int,matjson::Value> level_map;
 
-inline void RequestApi(bool retry,std::string Url, std::function<void(matjson::Value)> fun) {
+inline void RequestApi(bool retry,std::string Url, std::function<void(matjson::Value)> fun,std::function<void()> onfail) {
       web::WebRequest().get(Url).listen(
         [=](auto getval) {
             auto json = getval->json().unwrapOr("failed");
             if (json == "failed") {
                 //log::error("Failed at {}",Url);
                 if (retry) {
-                     RequestApi(retry,Url,fun);
-                } 
+                     RequestApi(retry,Url,fun,onfail);
+                } else {
+                    if (!onfail) {
+                        return;
+                    } else {
+                        try {
+                            onfail();
+                        } catch (const std::exception& ex) {
+                        log::error("Error doing function", ex.what());
+                    }
+                    }
+                }
                return;
             }
 
@@ -30,19 +40,32 @@ inline void RequestApi(bool retry,std::string Url, std::function<void(matjson::V
         [](auto prog) {
             
         },
-        []() {
-            log::warn("Request was cancelled.");
+        [=]() {
+             if (retry) {
+                     RequestApi(retry,Url,fun,onfail);
+                } else {
+                    if (!onfail) {
+                        return;
+                    } else {
+                        try {
+                            onfail();
+                        } catch (const std::exception& ex) {
+                        log::error("Error doing function", ex.what());
+                    }
+                    }
+                }
+            //log::warn("Request was cancelled.");
         }
     );
 }
 
-inline static void getlistjson(std::function<void(matjson::Value)> fun) {
+inline static void getlistjson(std::function<void(matjson::Value)> fun,std::function<void()> failed) {
    //log::debug("should send https://br-list.pages.dev/data/_list.json");
-    RequestApi(false,"https://br-list.pages.dev/data/_list.json", fun);
+    RequestApi(false,"https://br-list.pages.dev/data/_list.json", fun,failed);
 }
 
-inline static void getpackjson(std::function<void(matjson::Value)> fun) {
-    RequestApi(false,"https://br-list.pages.dev/data/_packlist.json", fun);
+inline static void getpackjson(std::function<void(matjson::Value)> fun,std::function<void()> failed) {
+    RequestApi(false,"https://br-list.pages.dev/data/_packlist.json", fun,failed);
 }
 
 inline static void getleveljson(const std::string& name, std::function<void(matjson::Value)> fun) {
@@ -55,5 +78,5 @@ inline static void getleveljson(const std::string& name, std::function<void(matj
         }
     }
     //log::debug("getting info for {}",name);
-    RequestApi(true,"https://br-list.pages.dev/data/" + finalurl + ".json", fun);
+    RequestApi(true,"https://br-list.pages.dev/data/" + finalurl + ".json", fun,nullptr);
 }