+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.1.0
+- Changed demon to a more fitting one
+- some changes on rate limit
+- server error connection message
+- tag on mod json
+- some bug fixes without node ids
 # 1.0.0
 - Pacific demon
 - list button in LevelSearchLayer